dmxbuffer-set-range!: take only byte-blob as source
It seems overly complex to try to support all srfi-4 vector types as
source buffers for the dmxbuffer set procedures, so starting with this
patch I'm trying out standardizing on one kind of source object.  In
this patch, we use byte-blob (from the byte-blob egg), but this may
change in the future, after getting a better sense of the advantages and
disadvantages of the available blob types.