init-logging: binding to InitLogging
Binding the log_output type proved tricky, because it is a typedef enum
and also because InitLogging is overloaded.  When we tried to pass our
enum type as an integer to InitLogging, C++ gave an ambiguity error.
Our solution (thanks zbigniew!) was to foreign-declare a wrapper
function for InitLogging, and have our binding call that wrapper.